+++ laser_ws/src/laser_package/src/drivers/track_driver.cpp
@@ -5,7 +5,7 @@ Tracker::Tracker()
 	ROS_INFO("Empty tracker");
 }
 
-Tracker::Tracker(Eigen::MatrixXf an_initial_state, float a_sampling_interval, Eigen::MatrixXf noise_data, int CT_model)
+Tracker::Tracker(Eigen::MatrixXd an_initial_state, double a_sampling_interval, Eigen::MatrixXd noise_data, int CT_model)
 {
 	T = a_sampling_interval;
 	
@@ -40,7 +40,7 @@ Tracker::Tracker(Eigen::MatrixXf an_initial_state, float a_sampling_interval, Ei
 	//printValues();//USE TO DEBUG VALUES IN MATRICES
 }
 
-void Tracker::initializeNoises(Eigen::MatrixXf noise_data)
+void Tracker::initializeNoises(Eigen::MatrixXd noise_data)
 {
 	mu_w = noise_data(MU_W,0);
 	sigma_w = noise_data(SIGMA_W,0);
@@ -96,11 +96,11 @@ void Tracker::initializeMatrices()
 
 void Tracker::updateCoordinatedTurnJacobian()
 {
-	float xi_hat = x_hat(XI);
-	float xi_dot_hat = x_hat(XI_DOT);
-	float eta_hat = x_hat(ETA);
-	float eta_dot_hat = x_hat(ETA_DOT);
-	float omega_hat = x_hat(OMEGA);
+	double xi_hat = x_hat(XI);
+	double xi_dot_hat = x_hat(XI_DOT);
+	double eta_hat = x_hat(ETA);
+	double eta_dot_hat = x_hat(ETA_DOT);
+	double omega_hat = x_hat(OMEGA);
 	if(fabs(omega_hat)>0.04)
 	{
 		updateOmegaPartials(xi_hat, xi_dot_hat, eta_hat, eta_dot_hat, omega_hat);
@@ -127,8 +127,8 @@ void Tracker::updateCoordinatedTurnJacobian()
 
 void Tracker::initializeCoordinatedTurnSystemMatrix()
 {
-	float omega = x_hat(OMEGA);
-	float inverse_omega = 1/omega; 
+	double omega = x_hat(OMEGA);
+	double inverse_omega = 1/omega; 
 	F <<
 	1, sin(omega*T)*inverse_omega, 0, (1-cos(omega*T))*inverse_omega, 0,
 	0, cos(omega*T), 0, -sin(omega*T), 0,
@@ -147,7 +147,7 @@ void Tracker::initializeUniformMotionSystemMatrix()
 	0, 0, 0, 0, 0;
 }
 
-void Tracker::update(Eigen::Vector2f z, double an_update_time, int CT_model)
+void Tracker::update(Eigen::Vector2d z, double an_update_time, int CT_model)
 {
 	updateDerivatives(z, an_update_time);
 	
@@ -173,15 +173,15 @@ void Tracker::update(Eigen::Vector2f z, double an_update_time, int CT_model)
 }
 
 
-void Tracker::updateDerivatives(Eigen::Vector2f z, double time_of_measurement)
+void Tracker::updateDerivatives(Eigen::Vector2d z, double time_of_measurement)
 {
-	//float speed_temp_sum = 0.0;
-	float some_x = z(0);
-	float some_y = z(1);
-	float velocity_x_temp_sum = 0.0;
-	float velocity_y_temp_sum = 0.0;
-	float accel_x_temp_sum = 0.0;
-	float accel_y_temp_sum = 0.0;
+	//double speed_temp_sum = 0.0;
+	double some_x = z(0);
+	double some_y = z(1);
+	double velocity_x_temp_sum = 0.0;
+	double velocity_y_temp_sum = 0.0;
+	double accel_x_temp_sum = 0.0;
+	double accel_y_temp_sum = 0.0;
 	for(int i = MEASUREMENT_MEMORY-1;i>0;i--)
 	{
 		//speed_temp_sum += speed_memory[i];
@@ -196,14 +196,14 @@ void Tracker::updateDerivatives(Eigen::Vector2f z, double time_of_measurement)
 		accel_x_memory[i] = accel_x_memory[i-1];
 		accel_y_memory[i] = accel_y_memory[i-1];
 	}
-	//float distanceBetweenMeasurements = getDistance(some_x, some_y, last_x, last_y);
-	float difference_of_x = some_x - last_x;
-	float difference_of_y = some_y - last_y;
-	float time_between_measurements = time_of_measurement - last_time;
-	float time_between_velocity_measurements = time_of_measurement - second_last_time;
-	float instantaneous_velocity_x = difference_of_x/time_between_measurements;
-	float instantaneous_velocity_y = difference_of_y/time_between_measurements;
-	//float current_instantaneous_speed = distanceBetweenMeasurements/timeBetweenMeasurements;
+	//double distanceBetweenMeasurements = getDistance(some_x, some_y, last_x, last_y);
+	double difference_of_x = some_x - last_x;
+	double difference_of_y = some_y - last_y;
+	double time_between_measurements = time_of_measurement - last_time;
+	double time_between_velocity_measurements = time_of_measurement - second_last_time;
+	double instantaneous_velocity_x = difference_of_x/time_between_measurements;
+	double instantaneous_velocity_y = difference_of_y/time_between_measurements;
+	//double current_instantaneous_speed = distanceBetweenMeasurements/timeBetweenMeasurements;
 	//speed_memory[0] = current_instantaneous_speed;
 	//speed_temp_sum += current_instantaneous_speed;
 	velocity_x_memory[0] = instantaneous_velocity_x;
@@ -230,7 +230,7 @@ void Tracker::updateDerivatives(Eigen::Vector2f z, double time_of_measurement)
 	ROS_INFO("X_accel = %f\n Y_accel = %f\n time between = %f\n\n-----------------------------------------------------", current_x_accel, current_y_accel, time_between_measurements);
 }
 
-void Tracker::updateOmegaPartials(float xi_hat, float xi_dot_hat, float eta_hat, float eta_dot_hat, float omega_hat)
+void Tracker::updateOmegaPartials(double xi_hat, double xi_dot_hat, double eta_hat, double eta_dot_hat, double omega_hat)
 {	
 	omega_partials(0) = (cos(omega_hat*T)*T*xi_dot_hat/omega_hat)-(sin(omega_hat*T)*xi_dot_hat/(omega_hat*omega_hat))-(sin(omega_hat*T)*T*eta_dot_hat/omega_hat)-((-1+cos(omega_hat*T))*eta_dot_hat/(omega_hat*omega_hat));
 	omega_partials(1) = -(sin(omega_hat*T)*T*xi_dot_hat)-(cos(omega_hat*T))*T*eta_dot_hat;
@@ -239,27 +239,27 @@ void Tracker::updateOmegaPartials(float xi_hat, float xi_dot_hat, float eta_hat,
 	ROS_INFO("updating omega partials");
 }
 
-float Tracker::getDistance(float x_1, float y_1, float x_2, float y_2)
+double Tracker::getDistance(double x_1, double y_1, double x_2, double y_2)
 {
 	return sqrt(pow((x_1-x_2),2)+pow((y_1-y_2),2));
 }
 
-float Tracker::getXAcceleration()
+double Tracker::getXAcceleration()
 {
 	return current_x_accel;
 }
 
-float Tracker::getYAcceleration()
+double Tracker::getYAcceleration()
 {
 	return current_y_accel;
 }
 
-float Tracker::getXVelocity()
+double Tracker::getXVelocity()
 {
 	return current_x_velocity;
 }
 
-float Tracker::getYVelocity()
+double Tracker::getYVelocity()
 {
 	return current_y_velocity;
 }
@@ -271,95 +271,95 @@ void Tracker::printValues()
 	//ROS_INFO("X_initial = [%f; %f] \n", x_hat[0](0), x_hat[0](1));
 }
 
-float Tracker::getPredictedX()
+double Tracker::getPredictedX()
 {
 	return x_hat(XI,0);
 }
 
-float Tracker::getPredictedY()
+double Tracker::getPredictedY()
 {
 	return x_hat(ETA,0);
 }
 
-float Tracker::getPredictedXVel()
+double Tracker::getPredictedXVel()
 {
 	return x_hat(XI_DOT,0);
 }
 
-float Tracker::getPredictedYVel()
+double Tracker::getPredictedYVel()
 {
 	return x_hat(ETA_DOT,0);
 }
 
-float Tracker::getPredictedOmega()
+double Tracker::getPredictedOmega()
 {
 	return x_hat(OMEGA,0);
 }
 
-float Tracker::getPositionVarianceX()
+double Tracker::getPositionVarianceX()
 {
 	return P(XI,XI);
 }
 
-float Tracker::getPositionVarianceY()
+double Tracker::getPositionVarianceY()
 {
 	return P(ETA,ETA);
 }
 
-float Tracker::getVelocityVarianceX()
+double Tracker::getVelocityVarianceX()
 {
 	return P(XI_DOT,XI_DOT);
 }
 
-float Tracker::getVelocityVarianceY()
+double Tracker::getVelocityVarianceY()
 {
 	return P(ETA_DOT,ETA_DOT);
 }
 
-float Tracker::getOmegaVariance()
+double Tracker::getOmegaVariance()
 {
 	return P(OMEGA, OMEGA);
 }
 
-float Tracker::getPositionGainX()
+double Tracker::getPositionGainX()
 {
 	return W(XI,XI);
 }
 
 
 
-float Tracker::getPositionGainY()
+double Tracker::getPositionGainY()
 {
 	return W(2,1);
 }
 
-float Tracker::getVelocityGainX()
+double Tracker::getVelocityGainX()
 {
 	return W(1,0);
 }
 
-float Tracker::getVelocityGainY()
+double Tracker::getVelocityGainY()
 {
 	return W(3,1);
 }
 
 
-float Tracker::getInnovationX()
+double Tracker::getInnovationX()
 {
 	return nu(0,0);
 }
 
-float Tracker::getInnovationY()
+double Tracker::getInnovationY()
 {
 	return nu(1,0);
 }
 
-float Tracker::getPredictedMeasurementX()
+double Tracker::getPredictedMeasurementX()
 {
 	return z_hat(0);
 }
 
-float Tracker::getPredictedMeasurementY()
+double Tracker::getPredictedMeasurementY()
 {
 	return z_hat(1);
 }
